test_ml/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastapi-utils==0.8.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.8.2
mypy-extensions==1.0.0
numpy==2.2.3
openai==0.28.0
packaging==24.2
pandas==2.2.3
psutil==5.9.8
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
